Automatic type passing is now working from the array items into their output.
There unfortunately seems to be an issue with Promises though as they have trouble inferring the type upon return when the type option is present for a snychronous return via a map delegate. Passing a "typeof" when creating the Promise seems to resolve the issue but it's interesting that it otherwise infers correctly when there is no non-promise return type.