Added several Array methods such as concat, find, etc...
Also switched to the "serial" designation instead of "ordered". Added some notes about possible code duplications that should be cleaned up. Several of the methods are "terminating" in the sense that they don't return an AAArray but instead async complete the actions and then return an inspection value (such as via "some") directly in the same Promise chain.