Polyfilled Array flat with types.
To work on older versions of Node (eg. 10, 11ish) that I want to support I need to polyfill Array.flat but the trouble is that the types don't play very nicely and get shunted to "unknown" for some reason. So there is a lot of @ts-expect-error comments for perfectly valid code and the type that AAArray eventually returns from its version of flat *is* correct so at least that's working.