Fixed up parallel "find" to now account for potential different speeds in result returning.
Not entirely happy though as it will still wait for all find callbacks to finish (ie. Promise.all) even if a result exists, should look into Promise.race and resolve wrapping / delaying to speed things up so that we return *immediately* once a positive result is found. Serial version is unaffected obviously. Made a note about having to do the same thing for findIndex.

Also broke apart the find and findSerial tests, should do this with other parallel/serial tests so it's easier to pinpoint issues in the test structure.

Added linting to the test script just because.